requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.5
attrs==23.2.0
babel==2.17.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.7.0
blis==1.0.1
caer==2.0.8
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.20.0
cmake==3.29.3
colorama==0.4.6
comm==0.2.0
confection==0.1.5
contourpy==1.2.0
cycler==0.12.1
cymem==2.0.10
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
dlib @ file:///C:/Users/mayan/OneDrive/Desktop/summer%20proj/Dlib-library-Installation-main/dlib-master/dlib-master/dlib-19.24.1-cp311-cp311-win_amd64.whl#sha256=6f1a5ee167975d7952b28e0ce4495f1d9a77644761cf5720fb66d7c6188ae496
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
executing==2.0.1
fastapi==0.115.12
fastjsonschema==2.21.1
filelock==3.14.0
flatbuffers==24.3.25
fonttools==4.46.0
fqdn==1.5.1
fsspec==2024.3.1
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
grpcio==1.68.1
h11==0.16.0
h5py==3.12.1
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.31.2
idna==3.7
intel-openmp==2021.4.0
ipykernel==6.27.1
ipython==8.19.0
isoduration==20.11.0
itsdangerous==2.1.2
jax==0.4.28
jaxlib==0.4.28
jedi==0.19.1
Jinja2==3.1.4
jiter==0.9.0
joblib==1.4.2
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-lsp==2.2.6
jupyter_client==8.6.0
jupyter_core==5.5.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
keras==3.11.3
kiwisolver==1.4.5
langcodes==3.5.0
language_data==1.3.0
lark==1.2.2
libclang==18.1.1
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mediapipe==0.10.14
mistune==3.1.3
mkl==2021.4.0
ml_dtypes==0.5.1
MouseInfo==0.1.3
mpmath==1.3.0
murmurhash==1.0.11
mypy==1.10.0
mypy-extensions==1.0.0
namex==0.0.8
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.5.8
networkx==3.3
notebook==7.4.4
notebook_shim==0.2.4
numpy==1.26.4
openai==1.78.1
opencv-contrib-python==4.6.0.66
opencv-python==4.7.0.68
opt-einsum==3.3.0
optree==0.13.1
overrides==7.7.0
packaging==23.2
pafy==0.5.5
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.3
Pillow==10.1.0
pinecone-client==6.0.0
pinecone-plugin-interface==0.0.7
platformdirs==4.1.0
preshed==3.0.9
prometheus_client==0.22.1
prompt-toolkit==3.0.43
protobuf==6.32.0
psutil==5.9.7
pure-eval==0.2.2
py-cpuinfo==9.0.0
PyAutoGUI==0.9.54
pybboxes==0.1.6
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
PyGetWindow==0.0.9
Pygments==2.17.2
PyMsgBox==1.0.9
pyparsing==3.1.1
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==3.3.0
python-magic==0.4.27
python-multipart==0.0.20
pytweening==1.2.0
pytz==2024.1
pywin32==306
pywinpty==2.0.15
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.36.2
regex==2024.11.6
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==13.9.4
rpds-py==0.26.0
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.13.0
seaborn==0.13.2
Send2Trash==1.8.3
sentence-transformers==4.1.0
shapely==2.0.4
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
smmap==5.0.1
sniffio==1.3.1
sounddevice==0.4.7
soupsieve==2.7
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
starlette==0.46.2
sympy==1.13.1
tbb==2021.12.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
terminado==0.18.1
terminaltables==3.1.10
thinc==8.3.2
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tinycss2==1.4.0
tokenizers==0.21.1
torch==2.5.1
torchvision==0.20.1
tornado==6.4
tqdm==4.66.4
traitlets==5.14.0
transformers==4.51.3
typer==0.15.1
types-python-dateutil==2.9.0.20250708
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.15
uri-template==1.3.0
urllib3==2.2.1
uvicorn==0.34.2
wasabi==1.1.3
wcwidth==0.2.12
weasel==0.4.1
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.1
wrapt==1.17.0
youtube-dl==2021.12.17
